+++ R/evaluate-promise.R
@@ -1,14 +1,7 @@
 #' Evaluate a promise, capturing all types of output.
 #'
-#' These functions allow you to capture the side-effects of a function call
-#' including printed output, messages and warnings. They are used to evaluate
-#' code for [expect_output()], [expect_message()],
-#' [expect_warning()], and [expect_silent()].
-#'
-#' @param code Code to evaluate. This should be an unevaluated expression.
-#' @param print If `TRUE` and the result of evaluating `code` is
-#'   visible this will print the result, ensuring that the output of printing
-#'   the object is included in the overall output
+#' @param code Code to evaluate.
+#' @keywords internal
 #' @export
 #' @return A list containing
 #'  \item{result}{The result of the function}
@@ -61,63 +54,3 @@ evaluate_promise <- function(code, print = FALSE) {
   )
 }
 
-#' @export
-#' @rdname evaluate_promise
-capture_messages <- function(code) {
-  out <- Stack$new()
-
-  withCallingHandlers(
-    code,
-    message = function(condition) {
-      out$push(condition)
-      invokeRestart("muffleMessage")
-    }
-  )
-
-  get_messages(out$as_list())
-}
-
-#' @export
-#' @rdname evaluate_promise
-capture_warnings <- function(code) {
-  out <- Stack$new()
-
-  withCallingHandlers(
-    code,
-    warning = function(condition) {
-      out$push(condition)
-      invokeRestart("muffleWarning")
-    }
-  )
-
-  get_messages(out$as_list())
-}
-
-get_messages <- function(x) {
-  vapply(x, "[[", "message", FUN.VALUE = character(1))
-}
-
-#' @export
-#' @rdname evaluate_promise
-#' @param width Number of characters per line of output
-capture_output <- function(code, print = FALSE, width = 80) {
-  output <- capture_output_lines(code, print, width = width)
-  paste0(output, collapse = "\n")
-}
-
-#' @export
-#' @rdname evaluate_promise
-capture_output_lines <- function(code, print = FALSE, width = 80) {
-  temp <- file()
-  on.exit(close(temp), add = TRUE)
-
-  withr::local_options(list(width = width))
-  withr::local_envvar(list(RSTUDIO_CONSOLE_WIDTH = width))
-
-  result <- withr::with_output_sink(temp, withVisible(code))
-  if (result$visible && print) {
-    withr::with_output_sink(temp, print(result$value))
-  }
-
-  read_lines(temp)
-}